+++ Client/Containers/CommandTypes.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Client.MessageHandlers
+namespace Client.Containers
 {
     public static class CommandTypes
     {
