+++ Client/ImageSender/ImageSender.cs
@@ -1,4 +1,4 @@
-﻿using Client.MessageHandlers;
+﻿using Client.Containers;
 using System.Net.Sockets;
 using System.Text;
